requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
openai>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
youtube-transcript-api>=0.6.0
sqlalchemy>=2.0.0
aiofiles>=23.0.0
httpx>=0.25.0
requests>=2.31.0
